refactor(api): delete hollow runtime application re-export shims
Seven files in modules/runtime/application were pure re-exports that
misrepresented where behavior lives: cross-layer shims for
infrastructure/domain symbols, a cross-module re-export of the mcp
runtime credential service, a rename of DriverInstance to
DriverConnection, and a single-caller array wrapper. Callers now import
the defining module directly; the DriverConnection platform alias
lives only at the durable-object adapter that wrangler binds.

Also dedupe parseAttachmentIds, which was defined identically in
send-agent-session-events.service and start-runs.service.
