--- apps/api/src/modules/runtime/application/driver-instance-do.service.ts
@@ -1,1 +1,0 @@
-export { DriverInstance as DriverConnection } from "../infrastructure/driver-instance/do";
