--- apps/api/src/modules/runtime/application/runtime-driver-routes.service.ts
@@ -1,1 +1,0 @@
-export { getRuntimeDriverRoutePrefix } from "../domain/runtime-driver-routes";
