--- apps/api/src/modules/runtime/application/runtime-maintenance.service.ts
@@ -1,1 +1,0 @@
-export { runSandboxMaintenance } from "../infrastructure/runtime-subject-lifecycle/runtime-subject-maintenance.service";
