--- apps/api/src/modules/runtime/application/runtime-mcp-credential.service.ts
@@ -1,4 +1,0 @@
-export {
-  invalidateRuntimeCredential,
-  refreshRuntimeCredential,
-} from "../../mcp/application/mcp-runtime.service";
