--- apps/api/src/modules/runtime/application/runtime-session-lifecycle.service.ts
@@ -1,4 +1,0 @@
-export { stopDriverSession } from "../infrastructure/driver-session-stop.service";
-export { closeSandboxConversationSession } from "../infrastructure/sandbox-session.service";
-export { deleteSandboxBackupsForDir } from "../infrastructure/sandbox-backup.service";
-export { destroyDriverInstanceDurableObject } from "../infrastructure/driver-instance/client";
