--- apps/api/src/modules/runtime/application/session-runs/session-run-summary.service.ts
@@ -1,5 +1,0 @@
-export { getSessionRunSummariesByIds } from "../../infrastructure/session-runs/session-run-store.repository";
-export {
-  toSessionRunSummary,
-  type SessionRunRow,
-} from "../../infrastructure/session-runs/session-run-row.mapper";
